pyproject.toml
[project]
name = "blogspot-publishing"
version = "0.1.0"
description = "Utility to publish a blog to Blogspot"
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/blogspot_publishing"]

[dependency-groups]
dev = ["pytest>=8.0.0", "ruff>=0.3.0"]

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
